changed int to uint16_t for port varible in events struct, same thing for redisLibeventAttach method, and added casting to uint16_t when redisLibeventAttach method is called

+++ src/apps/common/hiredis_libevent2.c
@@ -49,7 +49,7 @@ struct redisLibeventEvents {
   struct event *rev, *wev;
   int rev_set, wev_set;
   char *ip;
-  int port;
+  uint16_t port;
   char *pwd;
   int db;
 };
@@ -213,7 +213,7 @@ void send_message_to_redis(redis_context_handle rch, const char *command, const 
 
 ///////////////////////// Attach /////////////////////////////////
 
-redis_context_handle redisLibeventAttach(struct event_base *base, char *ip0, int port0, char *pwd, int db) {
+redis_context_handle redisLibeventAttach(struct event_base *base, char *ip0, uint16_t port0, char *pwd, int db) {
 
   struct redisLibeventEvents *e = NULL;
   redisAsyncContext *ac = NULL;
@@ -225,7 +225,7 @@ redis_context_handle redisLibeventAttach(struct event_base *base, char *ip0, int
     strncpy(ip, "127.0.0.1", sizeof(ip));
   }
 
-  int port = DEFAULT_REDIS_PORT;
+  uint16_t port = DEFAULT_REDIS_PORT;
   if (port0 > 0) {
     port = port0;
   }
